archive/prototype-rest/services/face/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
face_recognition==1.3.0
numpy==2.2.6
pillow==11.3.0
